Finish lab 2
Note about H: let m = min(arr), A >= m. You only need to move a number if it is not in the sorted position. You can only move A if gcd(A, m) == m. If gcd(A, m) != m, then gcd(A, x) != m for all x > m. Thus if you can't swap A and m, you can't swap A with anything, meaning the number is stuck in the wrong position. If you can swap A and m, you can place A in any unfixed position using m as a proxy for swaps.